+++ src-js/src/api-models/document.ts
@@ -8,32 +8,43 @@
 import { ApiGeometry } from "./geometry";
 
 export const enum ApiRelationshipType {
+  Answer = "ANSWER",
   Child = "CHILD",
   ComplexFeatures = "COMPLEX_FEATURES",
   MergedCell = "MERGED_CELL",
   Value = "VALUE",
 }
 
-export interface ApiRelationship {
+interface WithIds {
   Ids: string[];
-  Type: ApiRelationshipType;
 }
 
-export interface ApiChildRelationship extends ApiRelationship {
+export interface ApiAnswerRelationship extends WithIds {
+  Type: ApiRelationshipType.Answer;
+}
+
+export interface ApiChildRelationship extends WithIds {
   Type: ApiRelationshipType.Child;
 }
 
-export interface ApiComplexFeaturesRelationship extends ApiRelationship {
+export interface ApiComplexFeaturesRelationship extends WithIds {
   Type: ApiRelationshipType.ComplexFeatures;
 }
 
-export interface ApiMergedCellRelationship extends ApiRelationship {
+export interface ApiMergedCellRelationship extends WithIds {
   Type: ApiRelationshipType.MergedCell;
 }
 
-export interface ApiValueRelationship extends ApiRelationship {
+export interface ApiValueRelationship extends WithIds {
   Type: ApiRelationshipType.Value;
 }
+
+export type ApiRelationship =
+  | ApiAnswerRelationship
+  | ApiChildRelationship
+  | ApiComplexFeaturesRelationship
+  | ApiMergedCellRelationship
+  | ApiValueRelationship;
 
 export const enum ApiBlockType {
   Cell = "CELL",
@@ -41,6 +52,8 @@ export const enum ApiBlockType {
   Line = "LINE",
   MergedCell = "MERGED_CELL",
   Page = "PAGE",
+  Query = "QUERY",
+  QueryResult = "QUERY_RESULT",
   SelectionElement = "SELECTION_ELEMENT",
   Table = "TABLE",
   Word = "WORD",
@@ -137,6 +150,39 @@ export interface ApiMergedCellBlock {
   RowSpan: number;
 }
 
+export interface ApiQueryBlock {
+  BlockType: ApiBlockType.Query;
+  readonly Id: string;
+  /**
+   * Page number for this query block
+   *
+   * When a query is applied to multiple pages, it generates several QUERY blocks in the result -
+   * each the 'CHILD' of one page and each with a Page number.
+   */
+  Page: number;
+  readonly Query: {
+    Alias?: string;
+    Text: string;
+  };
+  /**
+   * Relationship links
+   *
+   * Relationships on this block type seem to contain only ANSWERs, and whole field appears to be
+   * omitted when no anwers found on a given page.
+   */
+  Relationships?: ApiRelationship[];
+}
+
+export interface ApiQueryResultBlock {
+  BlockType: ApiBlockType.QueryResult;
+  Confidence: number;
+  Geometry?: ApiGeometry;
+  readonly Id: string;
+  Page: number;
+  Text: string;
+  SearchKey: string;
+}
+
 export const enum ApiSelectionStatus {
   Selected = "SELECTED",
   NotSelected = "NOT_SELECTED",
@@ -156,6 +202,8 @@ export type ApiBlock =
   | ApiLineBlock
   | ApiMergedCellBlock
   | ApiPageBlock
+  | ApiQueryBlock
+  | ApiQueryResultBlock
   | ApiSelectionElementBlock
   | ApiTableBlock
   | ApiWordBlock;